refactor. move containers to related folder
