+++ packages/adena-extension/src/pages/certify/change-network/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ChangeNetworkLayout from '@layouts/change-network-layout/change-network-layout';
-import ChangeNetworkContainer from '@containers/change-network-container/change-network-container';
+import ChangeNetworkContainer from './change-network-container';
 
 export default function ChangeNetwork(): JSX.Element {
   return <ChangeNetworkLayout changeNetwork={<ChangeNetworkContainer />} />;
