+++ packages/adena-extension/src/pages/wallet/add-custom-network/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import AddCustomNetworkLayout from '@layouts/add-custom-network-layout/add-custom-network-layout';
-import AddCustomNetworkContainer from '@containers/add-custom-network-conatiner/add-custom-network-conatiner';
+import AddCustomNetworkContainer from './add-custom-network-conatiner';
 
 export default function AddCustomNetworkPage(): JSX.Element {
   return <AddCustomNetworkLayout addCustomNetwork={<AddCustomNetworkContainer />} />;
