+++ packages/adena-extension/src/pages/wallet/approve-adding-network/approve-adding-network.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ApproveAddingNetworkLayout from '@layouts/approve-adding-network-layout/approve-adding-network-layout';
-import ApproveAddingNetworkContainer from '@containers/approve-adding-network-container/approve-adding-network-container';
+import ApproveAddingNetworkContainer from './approve-adding-network-container';
 
 export default function ApproveAddingNetworkPage(): JSX.Element {
   return <ApproveAddingNetworkLayout approveAddingNetwork={<ApproveAddingNetworkContainer />} />;
