+++ packages/adena-extension/src/pages/wallet/approve-changing-network/approve-changing-network.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ApproveChangingNetworkLayout from '@layouts/approve-changing-network-layout/approve-changing-network-layout';
-import ApproveChangingNetworkContainer from '@containers/approve-changing-network-container/approve-changing-network-container';
+import ApproveChangingNetworkContainer from './approve-changing-network-container';
 
 export default function ApproveChangingNetworkPage(): JSX.Element {
   return (
