+++ packages/adena-extension/src/pages/wallet/approve-establish/index.tsx
@@ -1,6 +1,6 @@
-import ApproveEstablishContainer from '@containers/approve-establish-container/approve-establish-container';
+import ApproveEstablishContainer from './approve-establish-container';
 import React from 'react';
 
 export const ApproveEstablish: React.FC = () => {
-  return <ApproveEstablishContainer />
+  return <ApproveEstablishContainer />;
 };
