+++ packages/adena-extension/src/pages/wallet/approve-sign-ledger-loading/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import ApproveSignLedgerLoadingContainer from '@containers/approve-sign-ledger-loading-container/approve-sign-ledger-loading-container';
+import ApproveSignLedgerLoadingContainer from './approve-sign-ledger-loading-container';
 
 export const ApproveSignLedgerLoading = (): JSX.Element => {
   return <ApproveSignLedgerLoadingContainer />;
