+++ packages/adena-extension/src/pages/wallet/approve-sign-transaction-ledger-loading/approve-sign-transaction-ledger-loading.tsx
