+++ packages/adena-extension/src/pages/wallet/approve-sign/index.tsx
@@ -1,4 +1,4 @@
-import ApproveSignContainer from '@containers/approve-sign-container/approve-sign-container';
+import ApproveSignContainer from './approve-sign-container';
 import React from 'react';
 
 export const ApproveSign = (): JSX.Element => {
