+++ packages/adena-extension/src/pages/wallet/approve-transaction-ledger-loading/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import ApproveTransactionLedgerLoadingContainer from '@containers/approve-transaction-ledger-loading-container/approve-transaction-ledger-loading-container';
+import ApproveTransactionLedgerLoadingContainer from './approve-transaction-ledger-loading-container';
 
 export const ApproveTransactionLedgerLoading = (): JSX.Element => {
   return <ApproveTransactionLedgerLoadingContainer />;
