+++ packages/adena-extension/src/pages/wallet/approve-transaction-main/index.tsx
@@ -1,4 +1,4 @@
-import ApproveTransactionContainer from '@containers/approve-transaction-container/approve-transaction-container';
+import ApproveTransactionContainer from './approve-transaction-container';
 import React from 'react';
 
 export const ApproveTransactionMain = (): JSX.Element => {
