+++ packages/adena-extension/src/pages/wallet/edit-custom-network/index.tsx
@@ -1,6 +1,6 @@
-import EditCustomNetworkContainer from '@containers/edit-custom-network-container/edit-custom-network-container';
-import EditCustomNetworkLayout from '@layouts/edit-custom-network-layout/edit-custom-network-layout';
 import React from 'react';
+import EditCustomNetworkLayout from '@layouts/edit-custom-network-layout/edit-custom-network-layout';
+import EditCustomNetworkContainer from './edit-custom-network-container';
 
 export default function EditCustomNetworkPage(): JSX.Element {
   return <EditCustomNetworkLayout editCustomNetwork={<EditCustomNetworkContainer />} />;
