+++ packages/adena-extension/src/pages/wallet/history/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import HistoryLayout from '@layouts/history-layout/history-layout';
-import HistoryContainer from '@containers/history-container/history-container';
+import HistoryContainer from './history-container';
 
 export const History = (): JSX.Element => {
   return <HistoryLayout history={<HistoryContainer />} />;
