+++ packages/adena-extension/src/pages/wallet/manage-token/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ManageTokenLayout from '@layouts/manage-token-layout/manage-token-layout';
-import ManageTokenSearchContainer from '@containers/manage-token-search-container/manage-token-search-container';
+import ManageTokenSearchContainer from './manage-token-search-container';
 
 export default function ManageToken(): JSX.Element {
   return <ManageTokenLayout manageTokenSearch={<ManageTokenSearchContainer />} />;
