+++ packages/adena-extension/src/pages/wallet/transfer-summary/index.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import TransferSummaryLayout from '@layouts/transfer-summary-layout/transfer-summary-layout';
+import TransferSummaryContainer from './transfer-summary-container';
+
+const TransferSummary: React.FC = () => {
+  return <TransferSummaryLayout transferSummary={<TransferSummaryContainer />} />;
+};
+
+export default TransferSummary;
